[ANDROSDK-995] Move note package outside of enrollment
