+++ core/src/main/java/org/hisp/dhis/android/core/enrollment/internal/EnrollmentHandler.java
@@ -38,9 +38,9 @@ import org.hisp.dhis.android.core.arch.handlers.internal.IdentifiableDataHandler
 import org.hisp.dhis.android.core.common.State;
 import org.hisp.dhis.android.core.enrollment.Enrollment;
 import org.hisp.dhis.android.core.enrollment.EnrollmentInternalAccessor;
-import org.hisp.dhis.android.core.enrollment.note.Note;
-import org.hisp.dhis.android.core.enrollment.note.internal.NoteDHISVersionManager;
-import org.hisp.dhis.android.core.enrollment.note.internal.NoteUniquenessManager;
+import org.hisp.dhis.android.core.note.Note;
+import org.hisp.dhis.android.core.note.internal.NoteDHISVersionManager;
+import org.hisp.dhis.android.core.note.internal.NoteUniquenessManager;
 import org.hisp.dhis.android.core.event.Event;
 
 import java.util.ArrayList;
