+++ core/src/main/java/org/hisp/dhis/android/core/note/internal/NoteDHISVersionManager.java
@@ -26,31 +26,53 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.enrollment.note;
+package org.hisp.dhis.android.core.note.internal;
 
-import org.hisp.dhis.android.core.common.BaseObjectShould;
-import org.hisp.dhis.android.core.common.ObjectShould;
-import org.junit.Test;
+import org.hisp.dhis.android.core.common.BaseIdentifiableObject;
+import org.hisp.dhis.android.core.enrollment.Enrollment;
+import org.hisp.dhis.android.core.note.Note;
+import org.hisp.dhis.android.core.systeminfo.DHISVersionManager;
 
-import java.io.IOException;
 import java.text.ParseException;
 
-import static org.assertj.core.api.Java6Assertions.assertThat;
+import javax.inject.Inject;
 
-public class Note30Should extends BaseObjectShould implements ObjectShould {
+import dagger.Reusable;
 
-    public Note30Should() {
-        super("enrollment/note/note_30.json");
+@Reusable
+public class NoteDHISVersionManager {
+
+    private final DHISVersionManager versionManager;
+
+    @Inject
+    public NoteDHISVersionManager(DHISVersionManager versionManager) {
+        this.versionManager = versionManager;
     }
 
-    @Override
-    @Test
-    public void map_from_json_string() throws IOException, ParseException {
-        Note note = objectMapper.readValue(jsonStream, Note.class);
+    public Note transform(Enrollment enrollment, Note note) {
+        Note.Builder builder = Note.builder().enrollment(enrollment.uid());
 
-        assertThat(note.uid()).isEqualTo("noteUid");
-        assertThat(note.value()).isEqualTo("Note");
-        assertThat(note.storedBy()).isEqualTo("android");
-        assertThat(note.storedDate()).isEqualTo("2018-03-19T15:20:55.058");
+        try {
+            if (this.versionManager.is2_29()) {
+                builder
+                        .storedDate(BaseIdentifiableObject.dateToDateStr(
+                        BaseIdentifiableObject.parseSpaceDate(note.storedDate())))
+                        .uid(null);
+            } else {
+                builder
+                        .storedDate(BaseIdentifiableObject.dateToDateStr(
+                        BaseIdentifiableObject.parseDate(note.storedDate())))
+                        .uid(note.uid());
+            }
+        } catch (ParseException ignored) {
+            builder
+                    .storedDate(null)
+                    .uid(null);
+        }
+
+        return builder
+                .value(note.value())
+                .storedBy(note.storedBy())
+                .build();
     }
 }
