+++ core/src/test/java/org/hisp/dhis/android/core/enrollment/internal/EnrollmentHandlerShould.java
@@ -36,9 +36,9 @@ import org.hisp.dhis.android.core.arch.handlers.internal.Transformer;
 import org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils;
 import org.hisp.dhis.android.core.enrollment.Enrollment;
 import org.hisp.dhis.android.core.enrollment.EnrollmentInternalAccessor;
-import org.hisp.dhis.android.core.enrollment.note.Note;
-import org.hisp.dhis.android.core.enrollment.note.internal.NoteDHISVersionManager;
-import org.hisp.dhis.android.core.enrollment.note.internal.NoteUniquenessManager;
+import org.hisp.dhis.android.core.note.Note;
+import org.hisp.dhis.android.core.note.internal.NoteDHISVersionManager;
+import org.hisp.dhis.android.core.note.internal.NoteUniquenessManager;
 import org.hisp.dhis.android.core.event.Event;
 import org.junit.Before;
 import org.junit.Test;
